<WIP> Alters publishing flow
Instead of, after creating a new document being taken to the overview edit
page. The user will now be redirected to the topic tagging page if the
documents is eligible to be tagged to topic taxons, then, after saving, the
legacy tagging page. If the document is not eligible, the user will be
redirected straight to the legacy tagging page. After then saving the legacy tags,
the publisher will land on the overview edit page.

This flow will only happen once, when the document is first created.
After that taxons can be tagged from the overview edit page only.